New Stock API, Remove Wokwi, Formate Config
Needed a new API that can handle more volume for launch. Original one was just intended to be a placeholder and is not scaleable. Wokwi wont really be needed anymore now that dev hardware is avail, it also added allot of bloat. Changed the config file to group all the need-to-configure items near the top.