Stock API switch to twelvedata + new styling
Pushing this to both dev and main(not norm to push right to main, but need to do this as a one-off as the previous stock API has stopped working).

API now supports global markets, crypto,  and forex. See default examples for how to use these.

A small thing to be fixed: the current default font does not support the euro or pound symbol for when showing non-NA stocks, so i just use "EUR/GBP" as a placeholder. Once we adress fonts this will be an easy thing to swap out.

Also removed some unused wokwi/display driver code in main now that we're all on the same hardware

Code is poorly commented & formated,. sorry lol