Added Webserver to configure/store config values
A Webserver allows the user to configure and store some of the config values to the flash using <Preferences.h>. 
This is Timezone, Weather Location, Stockticker List and Widget Cycle in Seconds (or 0 to disable). 
These values are then used in the Globaltime.cpp, Stockticker Widget and Weather Widget. 
If a configvalue is not in the flash, the config.h definition will be used. 
Use Chrome, Firefox or Edge to use the Website. Still IOS 18.1 cannot load the page, because there is a bug in Safari.